+++ front/lib/models/workspace.ts
@@ -1,4 +1,3 @@
-import assert from "assert";
 import {
   CreationOptional,
   DataTypes,
@@ -12,9 +11,6 @@ import {
 import { front_sequelize } from "@app/lib/databases";
 import { Subscription } from "@app/lib/models/plan";
 import { User } from "@app/lib/models/user";
-import { MemberAgentVisibilityType } from "@app/types/assistant/agent";
-
-import { AgentConfiguration } from "./assistant/agent";
 
 export class Workspace extends Model<
   InferAttributes<Workspace>,
@@ -124,84 +120,6 @@ Workspace.hasMany(Membership, {
   onDelete: "CASCADE",
 });
 Membership.belongsTo(Workspace);
-
-export class MemberAgentVisibility extends Model<
-  InferAttributes<MemberAgentVisibility>,
-  InferCreationAttributes<MemberAgentVisibility>
-> {
-  declare id: CreationOptional<number>;
-  declare createdAt: CreationOptional<Date>;
-  declare updatedAt: CreationOptional<Date>;
-
-  declare visibility: MemberAgentVisibilityType;
-
-  declare membershipId: ForeignKey<Membership["id"]>;
-  declare agentConfigurationId: ForeignKey<AgentConfiguration["id"]>;
-}
-
-MemberAgentVisibility.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      autoIncrement: true,
-      primaryKey: true,
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-      defaultValue: DataTypes.NOW,
-    },
-
-    visibility: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-  },
-  {
-    modelName: "member_agent_visibility",
-    sequelize: front_sequelize,
-    indexes: [
-      { fields: ["membershipId"] },
-      { fields: ["agentConfigurationId", "membershipId"], unique: true },
-    ],
-  }
-);
-
-Membership.hasMany(MemberAgentVisibility, {
-  foreignKey: { allowNull: false },
-  onDelete: "CASCADE",
-});
-AgentConfiguration.hasMany(MemberAgentVisibility, {
-  foreignKey: { allowNull: false },
-  onDelete: "CASCADE",
-});
-
-MemberAgentVisibility.addHook(
-  "beforeCreate",
-  "only_one_listing_for_private_agent",
-  async (memberAgentList) => {
-    const agentConfiguration = await AgentConfiguration.findOne({
-      where: {
-        id: memberAgentList.dataValues.agentConfigurationId,
-      },
-    });
-    if (!agentConfiguration)
-      throw new Error("Unexpected: Agent configuration not found");
-    if (agentConfiguration.scope === "private") {
-      const existingMemberAgentList = await MemberAgentVisibility.findOne({
-        where: {
-          agentConfigurationId: memberAgentList.dataValues.agentConfigurationId,
-        },
-      });
-      assert(!existingMemberAgentList);
-    }
-  }
-);
 
 export class MembershipInvitation extends Model<
   InferAttributes<MembershipInvitation>,
